acpi: Rename ACPI_PDC constants
ACPI_PDC constants prefix suggest that those constants are only relevant
in the context of the _PDC method. This is not true, as they can also be
used in _OSC context. Change prefix to more generic ACPI_PROC_CAP, that
better describe the purpose of those constants as they describe bits in
processor capabilities buffer. Rename pdc_intel.h to proc_cap_intel.h to
reflect the change in the prefix.

Suggested-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Michal Wilczynski <michal.wilczynski@intel.com>