ACPI: tiny-power-button: Replace acpi_driver with platform_driver
Tiny-power-button driver uses incorrectly struct acpi_driver to
register itself. This is wrong as the instances of the ACPI devices
are not meant to be literal devices, they're supposed to describe
ACPI entry of a particular device.

Use platform_driver instead of acpi_driver. In relevant places call
platform devices instances pdev to make a distinction with ACPI
devices instances.

Suggested-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Michal Wilczynski <michal.wilczynski@intel.com>