ACPI: SBS: Rename ACPI device instances from device to adev
Since transformation from ACPI driver to platform driver there are two
devices on which the driver operates - ACPI device and platform device.
For the sake of reader this calls for the distinction in their naming,
to avoid confusion. Rename relevant instances of ACPI device to adev,
as corresponding platform devices are called pdev.

Signed-off-by: Michal Wilczynski <michal.wilczynski@intel.com>